feat(login): adds error handling for the login async middleware

+++ controllers/library.js
@@ -1,27 +1,39 @@
 import { bookUser_db } from '../db/queries/library.js';
+import { asyncWrapper } from './middleware.js';
+import { CustomErr, errorMsg } from './validations.js';
 
-export const addToLibrary = async (req, res, next) => {
+export const addToLibrary = asyncWrapper(async (req, res, next) => {
     if (res.errors !== undefined) {
         return next();
     }
 
     const userData = req.user;
-    const bookId = req.params.bookId ?? res.book.id;
+    if (!userData) {
+        throw new CustomErr(errorMsg.missingParams, 404, 'missingUserData');
+    }
 
+    const bookId = req.params.bookId ?? res.book.id;
     if (!bookId) {
-        throw new Error('No book data to add into the library');
+        throw new CustomErr(errorMsg.library.noData, 404, 'missingUserData');
     }
 
     await bookUser_db.addBook(String(bookId), userData.id);
 
     next();
-};
+});
 
-export const removeFromLibrary = async (req, res, next) => {
+export const removeFromLibrary = asyncWrapper(async (req, _, next) => {
     const userData = req.user;
+    if (!userData) {
+        throw new CustomErr(errorMsg.missingParams, 404, 'missingUserData');
+    }
+
     const { bookId } = req.params;
+    if (!bookId) {
+        throw new CustomErr(errorMsg.library.noData, 404, 'missingUserData');
+    }
 
     await bookUser_db.removeBook(bookId, userData.id);
 
     next();
-};
+});
